requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
idna==2.6
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.14.3
Pillow==5.1.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
tqdm==4.23.3
urllib3==1.22
Werkzeug==0.14.1
